requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23

# Git analysis
gitpython==3.1.40

# AI and ML
torch==2.2.1
transformers==4.35.2
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3

# Visualization
plotly==5.18.0
networkx==3.2.1

# Code analysis
radon==6.0.1
pylint==3.0.2
coverage==7.3.2

# API integrations
slack-sdk==3.26.0
notion-client==2.1.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0